requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
datasets==2.14.4
dill==0.3.7
faiss-cpu==1.7.4
filelock==3.12.2
flask==2.3.3
Flask-RESTful==0.3.10
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gunicorn==21.2.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.28.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.159
urllib3==2.0.4
werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
